feat: refactoring
- rename the folder containing the graph queries
- remove the second level of depth
