+++ setup.cfg
@@ -3,5 +3,5 @@ package_dir =
     = src
 
 [options.package_data]
-read_namespace.read_github_namespace.graph =
+read_namespace.query_github =
     *.graphql
