+++ src/read_namespace/query_github/fetch_org_repos.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 from sgqlc.endpoint.http import HTTPEndpoint
-from ...utils.check_dict import nested_keys_exist
+from ..utils.check_dict import nested_keys_exist
 
 
 # TODO create a unit test ,
