+++ src/read_namespace/query_github/github-schema.graphql
